requirements.txt
fastapi
uvicorn
python-dotenv
redis
motor
nltk
huggingface-hub
openai
